intel-iommu: check definitions of CSRs and change the capability registers
Check and correct the definitions of all the CSRs in intel_iommu.c, including
the default values, bit semantics and so on.
Change the value of the Capability Register and Extended Capability Register
to make only the simplest functionality remain.
Add function define_quad_wo() for defining the write-only property of 64-bit.

Signed-off-by: Le Tan <tamlokveer@gmail.com>